refactored file system for tripcard + tripinfo
